refactor: addressing review comments
