+++ projects/common/src/telemetry/user-telemetry.service.ts
@@ -1,17 +1,17 @@
 import { Injectable } from '@angular/core';
 import { UserTraits } from './telemetry';
-import { UserTelemetryInternalService } from './user-telemetry-internal.service';
+import { UserTelemetryHelperService } from './user-telemetry-helper.service';
 
 @Injectable({ providedIn: 'root' })
 export class UserTelemetryService {
-  public constructor(private readonly userTelemetryInternalService: UserTelemetryInternalService) {}
+  public constructor(private readonly userTelemetryHelperService: UserTelemetryHelperService) {}
 
   public initialize(userTraits: UserTraits): void {
-    this.userTelemetryInternalService.initialize();
-    this.userTelemetryInternalService.identify(userTraits);
+    this.userTelemetryHelperService.initialize();
+    this.userTelemetryHelperService.identify(userTraits);
   }
 
   public shutdown(): void {
-    this.userTelemetryInternalService.shutdown();
+    this.userTelemetryHelperService.shutdown();
   }
 }
